==> 07_Active_community_analysis_V1.1.Rmd <==
---
title: "Zin SIP"
subtitle: "07 Analysis of active communities"
author: "Talia Gabay and Roey Angel"
date: "`r Sys.Date()`"
always_allow_html: true
output:
  rmarkdown::github_document:
    toc: true
    toc_depth: 5
    number_sections: false
    dev: "png"
    df_print: "kable"
    keep_html: true
---

```{r libraries, include=F, message=FALSE, warning = FALSE}
#Load packages
library(tidyverse) 
library(magrittr)
library(RColorBrewer) 
library(ggtext)
library(extrafont)
library(svglite) 
library(viridis) 
library(ggsci) 
library(ggvenn)
library(visdat) 
library(glue) 
library(patchwork)
library(vegan) 
library(phyloseq) 
library(speedyseq)  
library(Biostrings)
library(tibble)
```

```{r style settings, echo=F, message=F, warning=F, results="asis", cache=T}
options(width = 90, knitr.table.format = "html") 
cache.path <- "07_Active_community_analysis_cache/"
fig.path <- "07_Active_community_analysis_figures/" 
knitr::opts_chunk$set(  
  warning = FALSE,
  message = FALSE,
  cache = TRUE,
  dev = "png",
  # dev = c("svglite", "png"),
  dev.args = list(png = list(bg = 'white')),
  # dev.args = list(svglite = list(bg = 'white'), png = list(bg = 'white')),
  dpi = 600,
  #  fig.width = 12,
  #  fig.height = 8,
  cache.path = cache.path,
  fig.path = fig.path
)
```

```{r functions, include=F}
#' gz
#' Compress a file using gz and delete the uncompressed file
#'
#' @param in_path (Required) Path and file name of a file to be compressed
#'
#' @param out_path (Required) Path and file name of the compressed file
#'
#' @author Roey Angel
#'
#' @return none
#'
#' @usage gz(paste0(file.path, ".svg"), paste0(file.path, ".svgz"))
#' @export

gz <- function(in_path, out_path = tempfile()) {
  out <- gzfile(out_path, "w")
  writeLines(readLines(in_path), out)
  close(out)

  file.remove(in_path)
  invisible(out_path)
}

#' Save a plot to svgz and png
#' The function uses `svglite` and `agg_png` devices and then comresses the svg to an svgz
#'
#' @param filename (Required) Path and file name excluding suffix (Default: "Rplot")
#'
#' @param plot (Required) Plot to save, defaults to last plot displayed
#'
#' @param scale (Optional) Multiplicative scaling factor (Default: 1)
#' 
#' @param pwidth, @param pheight (Optional) Plot size in "in" (Defaults: 10, 8)
#'
#' @param dpi (Optional) Plot resolution (Default: 300)
#'
#' @author Roey Angel
#'
#' @return none
#'
#' @usage save_figure(filename = "Rplot", 
#'                        plot = last_plot(), 
#'                        pwidth = 10, 
#'                        pheight = 8, 
#'                        dpi = 300, 
#'                        scale = 1, 
#'                        bg = "white",
#'                        pointsize = 12)
#' @export

save_figure <- function(filename = "Rplot", 
                        plot = last_plot(), 
                        pwidth = 10, 
                        pheight = 8, 
                        dpi = 600, 
                        scale = 1, 
                        bg = "white",
                        pointsize = 12) {
  require(svglite)
  require(ragg)
  
  svglite(paste0(filename, ".svg"), 
          width = pwidth, 
          height = pheight)
  print(plot)
  invisible(dev.off())
  
  agg_png(paste0(filename, ".png"),
          width = pwidth, 
          height = pheight, 
          units = "in",
          res = dpi, 
          scaling = scale)
  print(plot)
  invisible(invisible(dev.off()))
  
  gz(paste0(filename, ".svg"), paste0(filename, ".svgz"))
} 

```
### Set parameters and load phyloseq object
```{r parameters, cache=T}
data_path <- "C:/Users/Talia Gabay/OneDrive - post.bgu.ac.il/PhD/SIP experiment/Analyses/Analysis_With_Roey/DADA2_pseudo/"
Proj_name <- "Zin_SIP"
Tree_file <- paste0(data_path, "Tree/DADA2_reps_seq_prev_filt.filtered.align.treefile")
#Load phyloseq object
Ps_file <- paste0(Proj_name, "_filt_wTree.Rds")
Ps_obj <- readRDS(file = paste0(data_path, Ps_file))

```
### subset phyloseq object
First subset phyloseq object created in _06_Diff_abund_DESeq2 to remove negative controls, light fractions and unlabeled samples
```{r subset ps object, cache=T}

Ps_obj %>% 
  subset_samples(Label..18O.== "Labelled" & Density.zone == "Heavy") -> ps_obj_only_labelled 

```
### Subsetting ASVs to include only active ones
In this section, we will choose active sequences that appear in the table 'DESeq_res_SIP_each_rep_LFC_sig_df'
We do this for each sample separately, since in each sample, different sequences were identified
as active. 
```{r subsetting each sample, cache=T}

#For each samples - look at DESeq table and see which sequences are active in a 
#specific sample 
labelled_seqs_Na_2 <- c("Seq_57", "Seq_235", "Seq_460", "Seq_127", "Seq_623") #select active sequence from table for a spesific sample, in this case "Na_2"
ps_obj_only_labelled %>%
  subset_samples(Treatment == "Natural" & Comparison_pair == "2") %>% # select relevant samples from the ps object based on the treatment and comparison pair
  otu_table() %>% # takes the OTU table from the phyloseq object
  as(., "matrix") %>%
  t() -> mat_Na_2 #creates a matrix
mat_Na_2[!(rownames(mat_Na_2) %in% labelled_seqs_Na_2), ] = 0 # gives zero value for sequences that are not in the labelled list
Na2_summed <- as.matrix(rowSums(mat_Na_2)) #Sum fractions
colnames(Na2_summed) <- c("Na_2") #Rename variable to fit sample names in metadata

#Do the same for the rest of the samples
labelled_seqs_Na_3 <- c("Seq_248",
                        "Seq_29",
                        "Seq_250",
                        "Seq_100",
                        "Seq_1052",
                        "Seq_10",
                        "Seq_8030",
                        "Seq_3",
                        "Seq_19",
                        "Seq_215",
                        "Seq_8",
                        "Seq_31",
                        "Seq_254",
                        "Seq_46",
                        "Seq_1812",
                        "Seq_133",
                        "Seq_39",
                        "Seq_734",
                        "Seq_91",
                        "Seq_49",
                        "Seq_1365",
                        "Seq_175",
                        "Seq_979",
                        "Seq_32",
                        "Seq_108",
                        "Seq_95")
ps_obj_only_labelled %>%
  subset_samples(Treatment == "Natural" & Comparison_pair == "3") %>%
  otu_table() %>%
  as(., "matrix") %>%
  t() -> mat_Na_3
mat_Na_3[!(rownames(mat_Na_3) %in% labelled_seqs_Na_3), ] = 0
Na3_summed <- as.matrix(rowSums(mat_Na_3)) 
colnames(Na3_summed) <- c("Na_3") 


labelled_seqs_Na_4 <- c("Seq_1",
                        "Seq_9",
                        "Seq_3",
                        "Seq_11",
                        "Seq_5",
                        "Seq_139")
ps_obj_only_labelled %>%
  subset_samples(Treatment == "Natural" & Comparison_pair == "4") %>%
  otu_table() %>%
  as(., "matrix") %>%
  t() -> mat_Na_4
mat_Na_4[!(rownames(mat_Na_4) %in% labelled_seqs_Na_4), ] = 0
Na4_summed <- as.matrix(rowSums(mat_Na_4)) 
colnames(Na4_summed) <- c("Na_4")

labelled_seqs_Na_5 <- c("Seq_116")
ps_obj_only_labelled %>%
  subset_samples(Treatment == "Natural" & Comparison_pair == "5") %>%
  otu_table() %>%
  as(., "matrix") %>%
  t() -> mat_Na_5
mat_Na_5[!(rownames(mat_Na_5) %in% labelled_seqs_Na_5), ] = 0
Na5_summed <- as.matrix(rowSums(mat_Na_5)) 
colnames(Na5_summed) <- c("Na_7")

labelled_seqs_Res_1 <- c("Seq_97",
                         "Seq_4",
                         "Seq_42",
                         "Seq_19",
                         "Seq_20",
                         "Seq_1",
                         "Seq_269",
                         "Seq_148",
                         "Seq_9",
                         "Seq_124",
                         "Seq_30",
                         "Seq_187",
                         "Seq_661",
                         "Seq_411",
                         "Seq_562",
                         "Seq_659",
                         "Seq_1183",
                         "Seq_1808",
                         "Seq_7632")
ps_obj_only_labelled %>%
  subset_samples(Treatment == "Restored" & Comparison_pair == "1") %>%
  otu_table() %>%
  as(., "matrix") %>%
  t() -> mat_Res_1
mat_Res_1[!(rownames(mat_Res_1) %in% labelled_seqs_Res_1), ] = 0
Res1_summed <- as.matrix(rowSums(mat_Res_1)) 
colnames(Res1_summed) <- c("Res_3")


labelled_seqs_Res_2 <- c("Seq_122",
                         "Seq_63",
                         "Seq_4",
                         "Seq_275",
                         "Seq_68")
ps_obj_only_labelled %>%
  subset_samples(Treatment == "Restored" & Comparison_pair == "2") %>%
  otu_table() %>%
  as(., "matrix") %>%
  t() -> mat_Res_2
mat_Res_2[!(rownames(mat_Res_2) %in% labelled_seqs_Res_2), ] = 0
Res2_summed <- as.matrix(rowSums(mat_Res_2)) 
colnames(Res2_summed) <- c("Res_4")


labelled_seqs_Res_3 <- c("Seq_239",
                         "Seq_504",
                         "Seq_11",
                         "Seq_57",
                         "Seq_5")
ps_obj_only_labelled %>%
  subset_samples(Treatment == "Restored" & Comparison_pair == "3") %>%
  otu_table() %>%
  as(., "matrix") %>%
  t() -> mat_Res_3
mat_Res_3[!(rownames(mat_Res_3) %in% labelled_seqs_Res_3), ] = 0
Res3_summed <- as.matrix(rowSums(mat_Res_3)) 
colnames(Res3_summed) <- c("Res_5")

labelled_seqs_Res_4 <- c("Seq_24",
                         "Seq_3",
                         "Seq_67",
                         "Seq_73")
ps_obj_only_labelled %>%
  subset_samples(Treatment == "Restored" & Comparison_pair == "4") %>%
  otu_table() %>%
  as(., "matrix") %>%
  t() -> mat_Res_4
mat_Res_4[!(rownames(mat_Res_4) %in% labelled_seqs_Res_4), ] = 0
Res4_summed <- as.matrix(rowSums(mat_Res_4)) 
colnames(Res4_summed) <- c("Res_7")


labelled_seqs_Res_5 <- c("Seq_15",
                         "Seq_24",
                         "Seq_11",
                         "Seq_3",
                         "Seq_124",
                         "Seq_171",
                         "Seq_138",
                         "Seq_347",
                         "Seq_612",
                         "Seq_10",
                         "Seq_283",
                         "Seq_97",
                         "Seq_562",
                         "Seq_1221",
                         "Seq_9",
                         "Seq_4",
                         "Seq_34",
                         "Seq_16",
                         "Seq_1065",
                         "Seq_62",
                         "Seq_543",
                         "Seq_1616",
                         "Seq_52",
                         "Seq_1002",
                         "Seq_59",
                         "Seq_2396",
                         "Seq_37",
                         "Seq_1149",
                         "Seq_217",
                         "Seq_161",
                         "Seq_67",
                         "Seq_55",
                         "Seq_47",
                         "Seq_462",
                         "Seq_2260")
ps_obj_only_labelled %>%
  subset_samples(Treatment == "Restored" & Comparison_pair == "5") %>%
  otu_table() %>%
  as(., "matrix") %>%
  t() -> mat_Res_5
mat_Res_5[!(rownames(mat_Res_5) %in% labelled_seqs_Res_5), ] = 0
Res5_summed <- as.matrix(rowSums(mat_Res_5)) 
colnames(Res5_summed) <- c("Res_8")

```
### Make a new ASV table of only active sequences
```{r bind samples to new ASV, cache=T}

#Bind all of the tables for each sample together
Only_active_ASV_table <- cbind(Na2_summed, Na3_summed, Na4_summed, Na5_summed, Res1_summed,
                               Res2_summed, Res3_summed, Res4_summed, Res5_summed)

```
### Create a venn diagram
For visualizing which sequences appear in each treatment and which overlap
```{r venn diagram, cache=T}

#First, make df of only natural samples
natural_only <- cbind(Na2_summed, Na3_summed, Na4_summed, Na5_summed)
#sum abundances of all samples and change column name
natural_summed <- as.data.frame(rowSums(natural_only))
colnames(natural_summed) <- c("Natural")
#filter out zero values
natural_no_zeros <- filter(natural_summed, Natural > 0)
#Make a column of sequence names
natural_no_zeros <- tibble::rownames_to_column(natural_no_zeros, "Natural_seqs")
#Do the same for restored samples
restored_only <- cbind(Res1_summed, Res2_summed, Res3_summed, Res4_summed, Res5_summed)
restored_summed <- as.data.frame(rowSums(restored_only))
colnames(restored_summed) <- c("Restored")
restored_no_zeros <- filter(restored_summed, Restored > 0)
restored_no_zeros <- tibble::rownames_to_column(restored_no_zeros, "Restored_seqs")
# Make a list of Natural and Restored active seqs to compare fo venn diagram
Venn <- list(Natural_seqs = natural_no_zeros$Natural_seqs, Restored_seqs = restored_no_zeros$Restored_seqs)
#Make venn diagram
Venn_actice <- ggvenn(
  Venn, 
  fill_color = c("#0073C2FF", "#CD534CFF"),
  stroke_size = 0.5, set_name_size = 5
)
save_figure(paste0(fig.path, "Venn_active"), 
            Venn_actice, 
            pwidth = 10, 
            pheight = 8,
            dpi = 600)

knitr::include_graphics(paste0(fig.path, "Venn_active", ".png"))

```
#Make a consolidated metadata table
create a metadata table where all fractions are consolidated to fit the subsetted ASV table
```{r creat subsetted metadata table, cache=T}

ps_metadata <- as.data.frame(sample_data(ps_obj_only_labelled))
metadata_active = subset(ps_metadata, select = c(Sample, Treatment, Label..18O., Density.zone, Fraction.no.) )
metadata_active <- subset(metadata_active, Fraction.no. == '2')
#Remove sample number 1 (is an outlier)
metadata_active <- metadata_active[-c(1), ] 
metadata_active <- as.data.frame(metadata_active)
rownames(metadata_active) <- metadata_active$Sample
metadata_active <- metadata_active[ ,-c(1)]

```
### Create a new phyloseq object of only active ASVs
```{r creat subsetted ps object, cache=T}

#Build phyloseq object of only active ASVs
asvtable <- otu_table(Only_active_ASV_table,taxa_are_rows = TRUE)
taxtable <-tax_table(tax_table(Ps_obj))
metadata <- sample_data(metadata_active, sample)
#create phyloseq file
ps_active_only <- phyloseq(asvtable,taxtable,metadata)
#Add phylogenetic tree to phyloseq object
tree_active <- read_tree(Tree_file)
ps_active_only <- merge_phyloseq(ps_active_only,
                                 phy_tree(tree_active))
#Filter out ASVs that don't appear in any sample
ps_active_only_filtered = filter_taxa(ps_active_only, function(x) sum(x) > 1, TRUE)

```
##Relative abuncances of active community
```{r relative abundance active community, cache=T}
#Abundances at phylum level
active.phylum <- ps_active_only_filtered %>%
  tax_glom(taxrank = "Phylum") %>%                     # agglomerate at chosen taxonomic level
  transform_sample_counts(function(x) {x/sum(x)} ) %>% # Transform to rel. abundance
  psmelt()                                          # Melt to long format
active.agg <- aggregate(active.phylum$Abundance, FUN=mean, by=list(Treatment=active.phylum$Treatment,Phylum=active.phylum$Phylum))
active.phylum.filtered <- filter(active.agg,active.agg$x > 0.01)
write.csv(active.agg, file= "active phyla table.csv")

#Comparing relative abundance - Natural and Post-mining (Mann-Whitney)
actino <- filter(active.phylum, Phylum == "Actinobacteriota")
with(actino, shapiro.test(Abundance[Treatment == "Natural"]))
with(actino, shapiro.test(Abundance[Treatment == "Restored"]))
bartlett.test(Abundance ~ Treatment, data=actino)
wilcox.test(Abundance ~ Treatment, data=actino)

bacteroi <- filter(active.phylum, Phylum == "Bacteroidota")
with(bacteroi, shapiro.test(Abundance[Treatment == "Natural"]))
with(bacteroi, shapiro.test(Abundance[Treatment == "Restored"]))
bartlett.test(Abundance ~ Treatment, data=bacteroi)
wilcox.test(Abundance ~ Treatment, data=bacteroi)

cyano <- filter(active.phylum, Phylum == "Cyanobacteria")
with(cyano, shapiro.test(Abundance[Treatment == "Natural"]))
with(cyano, shapiro.test(Abundance[Treatment == "Restored"]))
bartlett.test(Abundance ~ Treatment, data=cyano)
wilcox.test(Abundance ~ Treatment, data=cyano)

firmi <- filter(active.phylum, Phylum == "Firmicutes")
with(firmi, shapiro.test(Abundance[Treatment == "Natural"]))
with(firmi, shapiro.test(Abundance[Treatment == "Restored"]))
bartlett.test(Abundance ~ Treatment, data=firmi)
wilcox.test(Abundance ~ Treatment, data=firmi)

gammati <- filter(active.phylum, Phylum == "Gemmatimonadota")
with(gammati, shapiro.test(Abundance[Treatment == "Natural"]))
with(gammati, shapiro.test(Abundance[Treatment == "Restored"]))
bartlett.test(Abundance ~ Treatment, data=gammati)
wilcox.test(Abundance ~ Treatment, data=gammati)

myxo <- filter(active.phylum, Phylum == "Myxococcota")
with(myxo, shapiro.test(Abundance[Treatment == "Natural"]))
with(myxo, shapiro.test(Abundance[Treatment == "Restored"]))
bartlett.test(Abundance ~ Treatment, data=myxo)
wilcox.test(Abundance ~ Treatment, data=myxo)

patesci <- filter(active.phylum, Phylum == "Patescibacteria")
with(patesci, shapiro.test(Abundance[Treatment == "Natural"]))
with(patesci, shapiro.test(Abundance[Treatment == "Restored"]))
bartlett.test(Abundance ~ Treatment, data=patesci)
wilcox.test(Abundance ~ Treatment, data=patesci)

proteo <- filter(active.phylum, Phylum == "Proteobacteria")
with(proteo, shapiro.test(Abundance[Treatment == "Natural"]))
with(proteo, shapiro.test(Abundance[Treatment == "Restored"]))
bartlett.test(Abundance ~ Treatment, data=proteo)
wilcox.test(Abundance ~ Treatment, data=proteo)

unclass <- filter(active.phylum, Phylum == "Unclassified")
with(unclass, shapiro.test(Abundance[Treatment == "Natural"]))
with(unclass, shapiro.test(Abundance[Treatment == "Restored"]))
bartlett.test(Abundance ~ Treatment, data=unclass)
wilcox.test(Abundance ~ Treatment, data=unclass)

#Bar plot - all phyla
cl <- colors(distinct = TRUE)
set.seed(15887)
colours <- sample(cl,18)
bp.active <- ggplot(active.agg, aes(x = Treatment, y = x, fill = Phylum)) + 
  geom_bar(stat = "identity", width = 0.4) + theme_light(base_size = 18) + 
  theme_classic() +
  theme(legend.position="bottom", legend.text=element_text(size=20), 
        axis.text.x=element_text(size=18),
        axis.text.y=element_text(size=18),) +
  labs(x="", y="Relative abundance (%)") + 
  scale_fill_manual(values=c("#6A3D9A","#A6CEE3", "#1F78B4", 
                        "#FFFF99", "#B2DF8A", "#33A02C", 
                       "#FB9A99", "#E31A1C", "#FDBF6F", 
                       "#FF7F00", "#CAB2D6")) 
                       

save_figure(paste0(fig.path, "Phyla_abundances_active_community"), 
            bp.active, 
            pwidth = 13, 
            pheight = 9,
            dpi = 600)

knitr::include_graphics(paste0(fig.path, "Phyla_abundances_active_community", ".png"))

```

### Beta diversity analysis
we will examine the active community using weighed Unifrac
```{r ordination active, cache=T}

#weighed Unifrac
active_unifrac <- UniFrac(ps_active_only_filtered, weighted=TRUE, normalized=TRUE, parallel=FALSE, fast=TRUE) #Calculate Unifrac
ordu = ordinate(ps_active_only_filtered, "PCoA", "unifrac", weighted=TRUE) #Ordinate
ord_unifrac <- plot_ordination(ps_active_only_filtered, ordu, color="Treatment") #Plot ordination

#Adonis
adonis2(active_unifrac ~ Treatment, method = "unifrac", data = as(sample_data(ps_active_only_filtered), "data.frame"), permutations = 999)

#PCoA plot with ggplot
Unifrac_ggp = ord_unifrac + 
  geom_point(size = 4, aes(colour = Treatment)) + 
  stat_ellipse(level=0.95, aes(color = Treatment, group = Treatment)) +
  scale_colour_manual(values = c("#00AFBB", "#FC4E07")) +
  theme_classic() +
  theme(axis.text.y = element_text(colour = "black", size = 12, face = "bold"), 
        axis.text.x = element_text(colour = "black", face = "bold", size = 12), 
        legend.text = element_text(size = 12, face ="bold", colour ="black"), 
        legend.position = "right", axis.title.y = element_text(face = "bold", size = 14), 
        axis.title.x = element_text(face = "bold", size = 14, colour = "black"), 
        legend.title = element_text(size = 14, colour = "black", face = "bold"), 
        panel.background = element_blank(), panel.border = element_rect(colour = "black", fill = NA, size = 1.2),
        legend.key=element_blank()) 
 
Unifrac_ggp
save_figure(paste0(fig.path, "unifrac_PCoA"), 
            Unifrac_ggp, 
            pwidth = 10, 
            pheight = 8,
            dpi = 600)

knitr::include_graphics(paste0(fig.path, "unifrac_PCoA", ".png"))

```

```{r colophon, eval=T}

sessioninfo::session_info() %>%
  details::details(
    summary = 'Current session info',
    open    = TRUE
 )

```
